// Dakali/src/api/generated/models/FieldGroupRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { FieldRequest } from './FieldRequest';
export type FieldGroupRequest = {
    id: number;
    name: string;
    sortOrder: number;
    fields: Array<FieldRequest>;
};